2 этап(iter 63)

+++ stats-service/stats-dto/src/main/java/ru/practicum/dto/ViewStatsDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class ViewStatsDto {
     private String app;
     private String uri;
-    private Long hits;
+    private Integer hits;
 }
